Fix warnings.

+++ crypto/pkcs12/p12_decr.c
@@ -69,8 +69,9 @@
  */
 
 unsigned char * PKCS12_pbe_crypt(X509_ALGOR *algor, const char *pass,
-	     int passlen, unsigned char *in, int inlen, unsigned char **data,
-	     int *datalen, int en_de)
+				 int passlen, unsigned char *in, int inlen,
+				 unsigned char **data, size_t *datalen,
+				 int en_de)
 {
 	unsigned char *out;
 	int outlen, i;
@@ -111,12 +112,13 @@ unsigned char * PKCS12_pbe_crypt(X509_ALGOR *algor, const char *pass,
  */
 
 void * PKCS12_item_decrypt_d2i(X509_ALGOR *algor, const ASN1_ITEM *it,
-	     const char *pass, int passlen, ASN1_OCTET_STRING *oct, int zbuf)
+			       const char *pass, int passlen,
+			       ASN1_OCTET_STRING *oct, int zbuf)
 {
 	unsigned char *out;
 	const unsigned char *p;
 	void *ret;
-	int outlen;
+	size_t outlen;
 
 	if (!PKCS12_pbe_crypt(algor, pass, passlen, oct->data, oct->length,
 			       &out, &outlen, 0)) {
@@ -147,9 +149,10 @@ void * PKCS12_item_decrypt_d2i(X509_ALGOR *algor, const ASN1_ITEM *it,
  * if zbuf set zero encoding.
  */
 
-ASN1_OCTET_STRING *PKCS12_item_i2d_encrypt(X509_ALGOR *algor, const ASN1_ITEM *it,
-				       const char *pass, int passlen,
-				       void *obj, int zbuf)
+ASN1_OCTET_STRING *PKCS12_item_i2d_encrypt(X509_ALGOR *algor,
+					   const ASN1_ITEM *it,
+					   const char *pass, int passlen,
+					   void *obj, int zbuf)
 {
 	ASN1_OCTET_STRING *oct;
 	unsigned char *in = NULL;
@@ -164,7 +167,7 @@ ASN1_OCTET_STRING *PKCS12_item_i2d_encrypt(X509_ALGOR *algor, const ASN1_ITEM *i
 		return NULL;
 	}
 	if (!PKCS12_pbe_crypt(algor, pass, passlen, in, inlen, &oct->data,
-				 &oct->length, 1)) {
+			      &oct->length, 1)) {
 		PKCS12err(PKCS12_F_PKCS12_ITEM_I2D_ENCRYPT,PKCS12_R_ENCRYPT_ERROR);
 		OPENSSL_free(in);
 		return NULL;
